feat(prepare): confidence-weighted fusion and cubic-spline gap filling
perception/geometry.py::lift_to_3d now emits a per-landmark fusion weight
w = visibility · sensor_validity · d^-2 · max(0, cosθ) (paper §3.5, eq. 2);
the incidence term comes from a 3-pixel depth-normal estimate and drops out on
a degenerate patch.

perception/recorder.py slimmed to the single rec.npz writer (write_rec), schema
== contracts.REC_KEYS with the confidence column. perception/extract.py feeds it
SkeletonFrames + per-landmark weights instead of building the npz inline.

prepare/run.py: pass the per-camera weights to fuse_trajectories; fill the fused
trajectory's remaining gaps with prepare.interpolate.fill_se3_spline (real cubic
spline, linear fallback < 4 valid, paper §3.7) before smoothing; ω_t is now the
mean over the wrist-frame landmarks of the max-over-cameras weight (eq. 5),
zeroed where the EE pose failed validation.

tests: +weighted-lift, +fill_se3_spline, + a --runslow real-IK episode test
(tests/conftest.py registers the marker and gates it). Suite: 101 passed / 1
skipped, 42 s.

+++ tests/unit_tests/e2e/test_pipeline_smoke.py
@@ -53,21 +53,18 @@ def test_pipeline_smoke(tmp_path):
         n_frames = len(d["positions"])
     assert n_frames > 0
 
-    # retarget is exercised separately (slow IK + model download); here we hand a
-    # synthetic plan.h5 to the replay leg so the artifact chain stays covered.
+    # retarget: the real PINK solve needs a network git-clone of the robot
+    # description on first run, so it is covered by tests/unit_tests/retarget/
+    # test_retarget_episode.py (skips cleanly offline) rather than here. Hand the
+    # replay leg a synthetic plan.h5 so the artifact chain stays exercised.
+    from viki.episode import mark_stage
     from viki.replay import replay_episode
     from viki.retarget.archive import write_hdf5_archive
 
     write_hdf5_archive(
         ep.plan_h5,
-        {
-            "q_scene_smooth": np.zeros((n_frames, 6), dtype=np.float64),
-            "dt": 1 / 30.0,
-            "robot": "",  # empty -> screen skips the model-load joint check (fast)
-        },
+        {"q_scene_smooth": np.zeros((n_frames, 6)), "dt": 1 / 30.0, "robot": ""},
     )
-    from viki.episode import mark_stage
-
     mark_stage(ep, "retarget", robot="")
 
     # replay: plan.h5 -> replay.h5 (dry-run)
